+++ src/lib/layer5_top_level/combinatorics/combinatorics_global.cpp
@@ -721,7 +721,7 @@ void combinatorics_global::Hill_cap56(
 										exit(1);
 										}
 									for (ii = 0; ii < sz[8]; ii++) {
-										int rk;
+										long int rk;
 										O->F->Projective_space_basic->PG_element_rank_modified(
 												pt_coords + set[ii] * n, 1, n, rk);
 										Sets[nb_sol * 56 + ii] = rk;
